Enable sass, add partial files and refactor styling
