+++ src/App.jsx
@@ -1,10 +1,10 @@
 import React, { useEffect, useState } from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import axios from "axios";
-import Header from "./components/Header";
+import Header from "./components/Header/Header";
 import Home from "./pages/Home";
 import Show from "./pages/Show/Show";
-import "./global.css";
+import "./global.scss";
 
 const App = () => {
   const [showsData, setShowsData] = useState([]);
