--- src/global.css
@@ -1,64 +1,0 @@
-/* CSS resets */
-
-html {
-  box-sizing: border-box;
-  font-size: 62.5%;
-}
-
-body {
-  font-size: 1.6rem;
-}
-
-body,
-h1,
-h2,
-h3,
-h4,
-h5,
-h6,
-p,
-ul,
-li {
-  margin: 0;
-}
-
-*,
-*::before,
-*::after {
-  box-sizing: inherit;
-}
-
-a {
-  text-decoration: none;
-  color: inherit;
-}
-
-ul {
-  list-style: circle;
-  padding: 0;
-}
-
-button {
-  border: none;
-}
-
-body {
-  font-family: Verdana, Geneva, Tahoma, sans-serif;
-  background-color: #ebeee9;
-}
-
-h1 {
-  font-size: 3.2rem;
-}
-
-.main {
-  display: flex;
-  justify-content: center;
-  margin: 1.6rem;
-  /* width: 100vw; */
-}
-
-.main section {
-  width: 100%;
-  max-width: 1200px;
-}
