+++ src/pages/Home.jsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import Card from "../components/Card";
+import Card from "../components/Card/Card";
 
 export const Home = ({ showsData }) => {
   const [showsCards, setShowsCards] = useState([]);
