+++ src/partials/_mixins.scss
@@ -1,0 +1,27 @@
+@use "./variables" as *;
+
+// Breakpoint mixins
+@mixin med {
+  @media (min-width: 768px) and (max-width: 1279px) {
+    @content;
+  }
+}
+
+@mixin lrg {
+  @media (min-width: 1280px) {
+    @content;
+  }
+}
+
+@mixin med-lrg {
+  @media (min-width: 768px) {
+    @content;
+  }
+}
+
+// Other mixins
+@mixin font-sizing($size, $height, $weight) {
+  font-size: $size;
+  line-height: $height;
+  font-weight: $weight;
+}
